+++ bin/miden-cli/src/commands/new_account.rs
@@ -281,7 +281,7 @@ async fn create_client_account<AUTH: TransactionAuthenticator + Sync + 'static>(
         builder = builder.with_component(component);
     }
 
-    let (account, seed) = builder
+    let account = builder
         .build()
         .map_err(|err| CliError::Account(err, "failed to build account".into()))?;
 
@@ -289,7 +289,7 @@ async fn create_client_account<AUTH: TransactionAuthenticator + Sync + 'static>(
         .add_key(&AuthSecretKey::RpoFalcon512(key_pair))
         .map_err(CliError::KeyStore)?;
 
-    client.add_account(&account, Some(seed), false).await?;
+    client.add_account(&account, false).await?;
 
     if deploy {
         deploy_account(client, &account).await?;