+++ crates/rust-client/src/lib.rs
@@ -136,13 +136,14 @@ pub mod builder;
 #[cfg(feature = "testing")]
 mod test_utils;
 
-mod errors;
+pub mod errors;
 
 // RE-EXPORTS
 // ================================================================================================
 
 /// Provides types and utilities for working with Miden Assembly.
 pub mod assembly {
+    pub use miden_objects::assembly::debuginfo::SourceManagerSync;
     pub use miden_objects::assembly::{
         Assembler,
         DefaultSourceManager,
@@ -166,6 +167,7 @@ pub mod asset {
     pub use miden_objects::asset::{
         Asset,
         AssetVault,
+        AssetWitness,
         FungibleAsset,
         NonFungibleAsset,
         NonFungibleAssetDetails,
@@ -242,6 +244,8 @@ pub mod testing {
 use alloc::sync::Arc;
 
 pub use miden_lib::utils::ScriptBuilder;
+use miden_objects::assembly::DefaultSourceManager;
+use miden_objects::assembly::debuginfo::SourceManagerSync;
 use miden_objects::block::BlockNumber;
 use miden_objects::crypto::rand::FeltRng;
 use miden_tx::LocalTransactionProver;
@@ -277,6 +281,8 @@ pub struct Client<AUTH> {
     /// An instance of a [`TransactionAuthenticator`] which will be used by the transaction
     /// executor whenever a signature is requested from within the VM.
     authenticator: Option<Arc<AUTH>>,
+    /// Shared source manager used to retain MASM source information for assembled programs.
+    source_manager: Arc<dyn SourceManagerSync>,
     /// Options that control the transaction executor’s runtime behaviour (e.g. debug mode).
     exec_options: ExecutionOptions,
     /// The number of blocks that are considered old enough to discard pending transactions.
@@ -332,12 +338,15 @@ where
             rpc_api.set_genesis_commitment(genesis.commitment()).await?;
         }
 
+        let source_manager: Arc<dyn SourceManagerSync> = Arc::new(DefaultSourceManager::default());
+
         Ok(Self {
             store,
             rng: ClientRng::new(rng),
             rpc_api,
             tx_prover,
             authenticator,
+            source_manager,
             exec_options,
             tx_graceful_blocks,
             max_block_number_delta,
@@ -351,7 +360,7 @@ where
 
     /// Returns an instance of the `ScriptBuilder`
     pub fn script_builder(&self) -> ScriptBuilder {
-        ScriptBuilder::new(self.in_debug_mode())
+        ScriptBuilder::with_source_manager(self.source_manager.clone())
     }
 
     /// Returns a reference to the client's random number generator. This can be used to generate