+++ crates/rust-client/src/test_utils/mock.rs
@@ -1,3 +1,4 @@
+use alloc::boxed::Box;
 use alloc::collections::{BTreeMap, BTreeSet};
 use alloc::sync::Arc;
 use alloc::vec::Vec;
@@ -211,13 +212,32 @@ impl MockRpcApi {
         self.mock_chain.read().committed_notes().values().cloned().collect()
     }
 
+    pub fn get_public_available_notes(&self) -> Vec<MockChainNote> {
+        self.mock_chain
+            .read()
+            .committed_notes()
+            .values()
+            .filter(|n| matches!(n, MockChainNote::Public(_, _)))
+            .cloned()
+            .collect()
+    }
+
+    pub fn get_private_available_notes(&self) -> Vec<MockChainNote> {
+        self.mock_chain
+            .read()
+            .committed_notes()
+            .values()
+            .filter(|n| matches!(n, MockChainNote::Private(_, _, _)))
+            .cloned()
+            .collect()
+    }
+
     pub fn advance_blocks(&self, num_blocks: u32) {
         let current_height = self.get_chain_tip_block_num();
         let mut mock_chain = self.mock_chain.write();
         mock_chain.prove_until_block(current_height + num_blocks).unwrap();
     }
 }
-use alloc::boxed::Box;
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 impl NodeRpcClient for MockRpcApi {
@@ -345,9 +365,9 @@ impl NodeRpcClient for MockRpcApi {
             .unwrap();
 
         if let Ok(account) = self.mock_chain.read().committed_account(account_id) {
-            Ok(FetchedAccount::Public(account.clone(), summary))
+            Ok(FetchedAccount::new_public(account.clone(), summary))
         } else {
-            Ok(FetchedAccount::Private(account_id, summary))
+            Ok(FetchedAccount::new_private(account_id, summary))
         }
     }